refactor: share update details and isolate update preferences
